### 2017-02-12 #181 @whedon
[PRE REVIEW]: SHGYield
**Submitting author:** @roguephysicist (<a href="http://orcid.org/0000-0002-4809-921X">Sean M. Anderson</a>)
**Repository:** <a href="https://github.com/roguephysicist/SHGYield" target ="_blank">https://github.com/roguephysicist/SHGYield</a>
**Version:** v1.0.1

Thanks for submitting your paper to JOSS roguephysicist. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
