### 2017-04-17 3598c63
added most recent publication to readme
