### 2017-04-17 #181 @roguephysicist
Hi @kyleniemeyer, I think that both of them seem like good candidates. Perhaps also @nicoguaro?

Thanks.
