### 2017-04-17 #181 @nicoguaro
@kyleniemeyer, I think I can review it. I like the plots in the repo ... so we have a nice beginning.
