### 2017-04-17 #242 @whedon
[REVIEW]: SHGYield
**Submitting author:** roguephysicist (<a href="http://orcid.org/0000-0002-4809-921X">Sean M. Anderson</a>)
**Repository:** <a href="https://github.com/roguephysicist/SHGYield" target ="_blank">https://github.com/roguephysicist/SHGYield</a>
**Version:** v1.0.1

- [x] **Authorship:** Has the submitting author (roguephysicist) made major contributions to the software?
