### 2017-04-17 #242 @roguephysicist
Thanks @nicoguaro and @kyleniemeyer. I am available for any questions you might have, though the theory behind the program is completely described in the references listed in the README (particularly [Phys. Rev. B 94, 115314 (2016)](https://doi.org/10.1103/PhysRevB.94.115314)). The most recent publication using this software has been accepted but is still undergoing final processing. You can find a [provisional version here](http://journal.frontiersin.org/article/10.3389/fmats.2017.00012/abstract).
