### 2017-04-17 #242 @nicoguaro
As a first comment, I would say that the main file should have docstrings to understand better the functions that appear there. As a second one, a description of the input and output files is missing, what makes really difficult to use the software for other cases different to the provided example.
