### 2017-04-18 #242 @roguephysicist
I can definitely see that. I will improve the documentation in the main file and add documentation to the  input file.
