### 2017-04-20 2a4aec1
added documentation to main file and input file
