+++ shgyield.py
@@ -4,6 +4,8 @@ coefficient for semiconductor surfaces. It works in conjunction with the matrix
 elements calculated with ABINIT, an open source ab initio software, and TINIBA,
 our in-house optical calculation software.
 
+For a complete overview of the theory, see PRB 94, 115314 (2016).
+
 Tested with Anaconda Python 4.0.0.
 
 requirements:
@@ -19,34 +21,51 @@ import numpy as np
 from scipy import constants, ndimage
 
 
-#### Functions ####
-def parse_input(infile): # parses input file from command line
+def parse_input(infile):
+    '''
+    Parses the input file specified on the command line. Returns dictionary
+    with variables and values from the input file.
+    '''
     params = {}
     targetfile = open(infile)
     data = targetfile.readlines()
     for line in data:
-        if '#' in line:
-            line = line.split('#', 1)
+        if line.strip().startswith('#') or line == '\n':
+            pass
         else:
-            key, value = line.split(":")
+            key, value = line.partition('#')[0].split(':')
             params[key.strip()] = value.strip()
     targetfile.close()
     return params
 
-def broad(target, sigma): 
+def broad(target, sigma):
+    '''
+    A function for applying Gaussian broadening on the final output data.
+    '''
     value = sigma * 42.666666666
     data = ndimage.filters.gaussian_filter(target, value)
     return data
 
-def savefile(file, freq, val1, val2, val3, val4): # saves to file with header
+def savefile(file, freq, val1, val2, val3, val4):
+    '''
+    Saves specified dataset to file with the following columns:
+    Energy(1w)    R_{pP}    R_{pS}    R_{sP}    R_{sS}
+    '''
     data = np.column_stack((freq, val1, val2, val3, val4))
-    np.savetxt(file, data,
-            fmt=('%05.2f', '%.14e', '%.14e', '%.14e', '%.14e'),
-            delimiter='    ',
-            header='RiF 1e-20 (cm^2/W)\n1w(eV) RpP'+21*" "+\
-                   'RpS'+21*" "+'RsP'+21*" "+'RsS')
+    np.savetxt(file, data, fmt=('%05.2f', '%.14e', '%.14e', '%.14e', '%.14e'),
+               delimiter='    ',
+               header='RiF 1e-20 (cm^2/W)\n1w(eV) RpP'+21*" "+\
+                      'RpS'+21*" "+'RsP'+21*" "+'RsS')
 
-def epsilon(in_file, norm): # loads chi1 file, creates epsilon numpy array
+def epsilon(in_file, norm):
+    '''
+    Reads calculated chi1 file, that is organized as
+    Energy(1w) Re[chi_xx] Im[chi_xx] Re[chi_yy] Im[chi_yy] Re[chi_zz] Im[chi_zz].
+    Converts to epsilon = 1 + 4*pi*chi1, and normalized chi1 according to
+    PRB 92, 245308 (2015). Returns the averaged epsilon
[…]
