### 2017-04-20 #242 @roguephysicist
I have added many comments to both the main and input files. Hopefully that will make things a lot easier to understand.
