### 2017-04-22 #242 @nicoguaro
I still don't get the sample-data input files thing. For example

    ## chi1 & chi2 components. Can be path to files, or number (like 0).

How come this can be a single number like 0, or a whole table like in the file ``sample-data/SiBulk-chi1-xx_yy_zz``.

I understand that you mention this

>You must first calculate the different components of the nonlinear susceptibility tensor. The theory surrounding this problem is still being developed, and there are many ways to go about it. We leave it to the reader to find the best method for their particular problem.

But let's suppose that I can "magically" know this tensor (that should have 27(?) components for every energy value). I still don't know how to format that information into inputs for the whole process.
