### 2017-04-24 #242 @roguephysicist
Some basic theory on the susceptibility tensor
----------------------------------------------

chi2 (chi1) is the nonlinear (linear) susceptibility tensor.

In particular, chi2 determines the nonlinear polarizability of a material and is responsible for second-harmonic generation. This relationship is expressed as
where **a**, **b**, and **c** are crystallographic directions that depend on how you orient your crystalline structure. We can see that a material can produce a polarization response in direction **a** from two incident fields (E) in directions **b** and **c**, by means of chi2^{**abc**}.

chi2 is a second-order tensor, and thus has 27 possible components (unique combinations of **a**, **b**, and **c**; for instance, **aaa**, **aab**, and so on.). Second-harmonic generation implies that the incoming fields are identical (two photons of equal energy in, one photon of double-energy out) so it is also implied that
for this particular phenomenon. This reduces 9 of the possible combinations, reducing to 18 unique components. It is very convenient to express the crystallographic directions in terms of x, y, and z; therefore, we can express chi2 with all 18 components as
Symmetry relations are very important for determining chi2. A given crystal symmetry can greatly reduce the complexity of the problem by eliminating many of the components. For instance, for the (001) face of cubic crystals, we have that
which has only 3 independent components. There are many [articles](https://journals.aps.org/prb/abstract/10.1103/PhysRevB.35.1129) and [books](https://books.google.com/books?id=bdFju3af2FsC) with tables and extensive discussion that describe the form that this tensor should have for a given symmetry.

As far as the program is concerned, you can indicate in the input file that you have all 18 components for your structure (maybe the symmetry of your material is unknown, so you calculate all of them), or that some of them are 0 (the program assumes that any of them NOT listed are 0).

The case for chi1 is considerable simpler. chi1 is directly related to the dielectric function of the material
which is directly related to the index of refraction as
The chi1 spectra should obviously have non-zero regions; otherwise, the problem is not very interesting.
