### 2017-04-24 #242 @roguephysicist
Concerning the calculation of the nonlinear susceptibility tensor
=================================================================

I think that the other part of your question concerns how we actually calculate chi2. That is definitely a non-trivial problem, and as I mentioned in the README, there are different ways to approach it. The method that we use is derived in full detail [here](https://doi.org/10.1103/PhysRevB.91.075302), although that reference does not really touch upon the relevant computational aspects.

This response ended up a little long-winded, so the last section essentially summarizes what I think you are asking.

On calculating the electronic properties of semiconductors
----------------------------------------------------------

In general, there are several *ab initio* formalisms that can be used to obtain the wavefunctions/electron densities/energies of crystalline semiconductor materials. The most common are:

* all-electron methods (older, declining in popularity)
* density functional theory with the local-density approximation (DFT-LDA) (solid workhorse with known problems, **used in this work**)
* time-dependent density functional theory (TDDFT) (good for optical responses, but hard to implement)
* many-body perturbation theory (MBPT) (state-of-the-art, but extreme computational expense)

Each of these has their pros and cons that mainly relate to accuracy vs. difficulty of the theory vs. computer resources. There are [many free and open-source codes](http://psi-k.net/software/) available for download, that are under active development by thousands of researchers and groups.

On calculating the susceptibility tensors and optical properties
----------------------------------------------------------------

Once the initial electronic properties of the material are determined, we can then proceed to calculate the optical properties that include the linear and nonlinear responses. In general, nonlinear optics is now pretty well understood within the DFT-LDA (see ref. above and refs. within) and the [TDDFT](http://dx.doi.org/10.1063/1.2790014) frameworks. MBPT is currently (AFAIK) still at the linear optics level, but has the highest level of accuracy available. Of course, you can combine different methods to exploit the strengths that each has to offer.

Unfortunately, most groups do not publish the software codes used to calculate the susceptibility tensor. The code that we use is called [TINIBA](https://github.com/roguephysicist/tiniba); however, it is VERY unpolished and requires  a major overhaul before it could ever be "officially" released/published. That said, we are an active research group that uses this software every day to produce high-quality scientific work, and are constantly improving and adding features to it. Publishing TINIBA is definitely a long-term goal that we have.

There are, however, some codes available, such as [RT-SIESTA](http://monalisa.phys.washington.edu/feffproject-rtsiesta.html) which works with TDDFT, and even [ABINIT](http://www.abinit.org/doc/helpfiles/for-v8.2/users/optic_help.html) has built-in utilities for calculating the susceptibility tensors.

On calculating the SHG yield
----------------------------

In summary, SHGYield.py produces the SHG radiation (in reflectance) produced from a semiconductor surface. As you mentioned, it requires the susceptibility tensors that are calculated with the above considerations. The theory is developed considering a reflectance model with three distinct regions that allows the user to readily simulate the SHG response of thin-films over bulk substrates, or of any crystalline surface with any symmetry considerations (see my previous reply to this thread). The program does NOT care how you produced the susceptibility tensors; you can use MBPT for producing the linear susceptibility tensor and then use TDDFT to produce the chi2 components, etc.

This is why I think that this program has value and is worth being published - it tackles a research problem (the SHG radiation from a surface), but allows the researcher to use their data that can be produced using a variety of frameworks and theoretical developments.
