### 2017-04-24 #242 @nicoguaro
@roguephysicist, these are good explanations, and I think that you should include them in the documentation of your software. I little detail, though, I suppose that you meant _third-order_ tensor, since chi2^{abc} has three free indices. On the other hand, chi1^{ab} is a second-order tensor, since it has two free indices.

So, you answered part of my question. But some part is still missing, and is the formatting of these files. And that was my main concern, because right now is not clear on how to format the information of the tensors so that the program read the properly.

> This is why I think that this program has value and is worth being published - it tackles a research problem (the SHG radiation from a surface), but allows the researcher to use their data that can be produced using a variety of frameworks and theoretical developments.

I also think that it is worth being published. But it is important that the documentation is good enough for people to use it.
