### 2017-04-24 #242 @roguephysicist
>  I suppose that you meant third-order tensor, since chi2^{abc} has three free indices. On the other hand, chi1^{ab} is a second-order tensor, since it has two free indices.

That was a
