### 2017-04-24 #242 @roguephysicist
> I suppose that you meant third-order tensor, since chi2^{abc} has three free indices. On the other hand, chi1^{ab} is a second-order tensor, since it has two free indices.

I thought one thing and wrote another! chi2 is a *second-order* nonlinear susceptibility, but is a *third-rank* tensor. Thanks for spotting that.

> So, you answered part of my question. But some part is still missing, and is the formatting of these files. And that was my main concern, because right now is not clear on how to format the information of the tensors so that the program read the properly.

Sorry for missing that. The chi2 components are dealt with in the `shgcomp()` function on line 76. The file is organized in 5 columns:

1. 1w energy from 0-20 eV in 0.01 eV steps,
2. real part of the 1w resonances
3. imaginary part of the 1w resonances
4. real part of the 2w resonances
5. imaginary part of the 2w resonances

We separate the 1w and 2w spectra because they are computed from separate equations, but in the end chi2^{abc} = Re[chi2(w) + chi2(2w] + i*Im[chi2(w) + chi2(2w] (line 96). The function also takes care of the scale and units that are produced in TINIBA (line 97), and can easily accommodate any set of units the user might want to use.

chi1 is handled in the `epsilon()` function. We calculate all three directions of chi1 (xx, yy, and zz) in the same file. Those are organized as:

1. 1w energy from 0-20 eV in 0.01 eV steps,
2. real part xx
3. imaginary part xx
4. real part yy
5. imaginary part yy
6. real part of zz
7. imaginary part of zz

chi1 gets averaged and converted to epsilon per the equations mentioned in my previous reply. This all happens on lines 70 -73.

> I also think that it is worth being published. But it is important that the documentation is good enough for people to use it.

You are exactly right, thanks. I will add some of this discussion to the README to really beef up the documentation.
