### 2017-04-28 e4fe3b6
Enhanced readme, changed input file to YAML, and improved functionality when interpreting the chi2 components
