+++ paper/paper.bib
@@ -6,6 +6,18 @@
   url    = {https://github.com/roguephysicist/SHGYield},
 }
 
+@article{andersonFMATS17,
+  author  = {Anderson, S. M. and Mendoza, B. S.},
+  title   = {Depth-Dependent Three-Layer Model for the Surface Second-Harmonic Generation Yield},
+  journal = {Frontiers in Materials},
+  volume  = {4},
+  pages   = {12},
+  year    = {2017},   
+  url     = {http://journal.frontiersin.org/article/10.3389/fmats.2017.00012},
+  issn    = {2296-8016},
+  doi     = {10.3389/fmats.2017.00012}
+}
+
 @article{andersonPRB16b,
   doi = {10.1103/PhysRevB.94.115314},
   year  = {2016},
