+++ shgyield.py
@@ -16,34 +16,25 @@ python shgyield.py <sample.in>
 """
 
 import sys
-import math
+import yaml
 import numpy as np
 from scipy import constants, ndimage
 
 
 def parse_input(infile):
     '''
-    Parses the input file specified on the command line. Returns dictionary
+    Parses the YAML input file specified on the command line. Returns dictionary
     with variables and values from the input file.
     '''
-    params = {}
-    targetfile = open(infile)
-    data = targetfile.readlines()
-    for line in data:
-        if line.strip().startswith('#') or line == '\n':
-            pass
-        else:
-            key, value = line.partition('#')[0].split(':')
-            params[key.strip()] = value.strip()
-    targetfile.close()
+    with open(infile) as data_file:
+        params = yaml.load(data_file)
     return params
 
 def broad(target, sigma):
     '''
     A function for applying Gaussian broadening on the final output data.
     '''
-    value = sigma * 42.666666666
-    data = ndimage.filters.gaussian_filter(target, value)
+    data = ndimage.filters.gaussian_filter(target, sigma)
     return data
 
 def savefile(file, freq, val1, val2, val3, val4):
@@ -52,10 +43,10 @@ def savefile(file, freq, val1, val2, val3, val4):
     Energy(1w)    R_{pP}    R_{pS}    R_{sP}    R_{sS}
     '''
     data = np.column_stack((freq, val1, val2, val3, val4))
-    np.savetxt(file, data, fmt=('%05.2f', '%.14e', '%.14e', '%.14e', '%.14e'),
+    np.savetxt(file, data, fmt=('%05.2f', '%.8e', '%.8e', '%.8e', '%.8e'),
                delimiter='    ',
-               header='RiF 1e-20 (cm^2/W)\n1w(eV) RpP'+21*" "+\
-                      'RpS'+21*" "+'RsP'+21*" "+'RsS')
+               header='RiF 1e-20 (cm^2/W)\n1w(eV) RpP'+15*" "+\
+                      'RpS'+15*" "+'RsP'+15*" "+'RsS')
 
 def epsilon(in_file, norm):
     '''
@@ -73,31 +64,15 @@ def epsilon(in_file, norm):
     epsa = 1 + (4 * constants.pi * coma * norm) # epsilon with normalization
     return epsa
 
-def shgcomp():
+def shgload(infile):
     '''
-    Reads each chi^{abc} listed in the input file, with the following columns:
-    Energy(1w) Re[1w] Im[1w] Re[2w] Im[2w].
-    Checks for all 18 components possible for SHG; those not listed in the input
-    file are assumed to be 0. Sums 1w and 2w real and imaginary parts, and
-    returns a dictionary with the component name as the 'key', and a numpy array
-    with the component data as the 'value'.
+    Loads chi^{abc} listed in the input file, with the following columns:
+    Energy(1w) Re[1w] Im[1w] Re[2w] Im[2w]. Sums 1w and 2w real and imaginary
[…]
