+++ README.md
@@ -3,21 +3,31 @@ SHG Yield for Semiconductor Surfaces
 
 [![DOI](https://zenodo.org/badge/11697217.svg)](https://zenodo.org/badge/latestdoi/11697217)
 
-SHGYield is a python program for calculating the second-harmonic generation (SHG) yield for semiconductor surfaces. This is useful for conducting theoretical second-harmonic spectroscopy of any crystalline semiconductor or nanostructure. The physical theory behind this phenomenon is presented in the references below. We have applied this software to several silicon test cases and have achieved very accurate results that compare well with experiment.
+SHGYield is a python program for calculating the surface second-harmonic generation (SSHG) yield (in reflectance) for semiconductor surfaces. 
 
-These calculations can predict effects in novel materials that are still under study, or save valuable time and resources for the experimentalist. For example, the figure below is an overview of the angular dependence of the reflected SHG Yield from the Si(111)(1x1)H surface. Experimentalists will find this very useful, as they can plan the experiment accordingly in order to optimize the output signal strength and polarization.
 
-![An overview of the angular dependence of the SHG Yield for the Si(111)(1x1)H surface](paper/3D-Si1x1.png)
+Introduction
+------------------------------------
 
-You must first calculate the different components of the nonlinear susceptibility tensor. The theory surrounding this problem is still being developed, and there are many ways to go about it. We leave it to the reader to find the best method for their particular problem. As an example, we use [ABINIT](http://www.abinit.org) to calculate the electron density and matrix elements and then [TINIBA](https://github.com/bemese/tiniba) to calculate the tensor components. 
+SSHG is an effective, nondestructive, and noninvasive probe for studying surface and interface properties, and even for characterizing buried interfaces and nanostructures. The high surface sensitivity of SSHG spectroscopy is due to the fact that within the dipole approximation, the bulk SHG response in centrosymmetric materials is identically zero. The SHG process can occur only at the surface where the inversion symmetry is broken. 
+
+This program has several potential applications and uses:
+* determining and analyzing the physical origin of SSHG spectra
+* predicting and characterizing the radiated SSHG for interesting new materials
+* characterizing thin films based on measured SH spectra
+* allowing the experimenter to calculate and analyze the SSHG yield to optimize experiments
+
+For example, the figure below is an overview of the angular dependence of the reflected SHG Yield from the Si(111)(1x1)H surface. Experimentalists will find this very useful, as they can plan the experiment accordingly in order to optimize the output signal strength and polarization.
+
+![An overview of the angular dependence of the SHG Yield for the Si(111)(1x1)H surface](paper/figures/3D-Si1x1.png)
 
 
 References
 ------------------------------------
 
-The complete theory is derived step-by-step in my [Ph.D. thesis](https://github.com/roguephysicist/thesis-phd). This software has been used in the following publications:
+The complete theory is derived step-by-step in [Phys. Rev. B 94, 115314 (2016)](https://doi.org/10.1103/PhysRevB.94.115314). This software has been developed and used in the following publications:
 
-* [Front. Mater. 4:12 (2017)](http://journal.frontiersin.org/article/10.3389/fmats.2017.00012/abstract)
+* [Front. Mater. 4:12 (2017)](https://doi.org/10.3389/fmats.2017.00012)
 * [Phys. Rev. B 94, 115314 (2016)](https://doi.org/10.1103/PhysRevB.94.115314)
 * [Phys. Rev. B 93, 235304 (2016)](https://doi.org/10.1103/PhysRevB.93.235304)
 * [Phys. Rev. B 91, 075302 (2015)](https://doi.org/10.1103/PhysRevB.91.075302)
[…]
