### 2017-04-28 #242 @roguephysicist
Hey @nicoguaro, I just updated the repo with the following changes.

I expanded the README with a lot of the information I posted above. I changed the input file to YAML which greatly improves the parsing abilities, and added a lot of comments to both the main file and the input file. I improved the way that the main file interprets the chi2 components, so it's smarter and more flexible now.

Let me know what you think.
