### 2017-04-28 #242 @nicoguaro
I think that it looks cleaner now, and the choice of YAML seems to be aligned in the same direction, as well.  It could be useful to do the same with the output, a Jupyter Notebook reproducing the plots from above might be good for that purpose.

Regarding automatic testing, it seems that it still does not cover this. The same goes for the community guidelines for contribution.
