### 2017-05-02 d13f71d
Added testing script, fleshing out example case
