+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+# How to contribute
+
+Thank you for using `SHGYield.py`; consider enhancing it with your valuable
+expertise! We welcome contributions, ideas for collaborations, and questions
+about `SHGYield.py`. You can contact us via the emails found in the [README
+file](README.md#license)).
+
+`SHGYield.py` is under constant development by an active and passionate research
+group. Check back frequently for new developments and published articles
+relating to this work.
