### 2017-05-02 #242 @roguephysicist
Thanks for these good suggestions.

I added a minimal test script in the `test` directory, and am working on a full fledged example with a Jupyter tutorial that includes a couple of different calculations. I'm still working on the Jupyter notebook (my first time), but most of the other stuff is already in place. I have no experience writing test programs, so I appreciate any feedback you might have. I also added a `CONTRIBUTING.md` file.

I have a few questions about the "JOSS paper" format. Should I enhance it with some of the stuff in the main README, or is it best to keep it simple? Also, does the LaTeX conversion happen automatically, or is that something I need to add in manually?
