### 2017-05-02 #242 @kyleniemeyer
@roguephysicist you are welcome to edit your paper, especially if @nicoguaro thinks it could use more content. I took a quick look, and agree it could use a bit more. Please look at some examples of published JOSS papers.

The LaTeX conversion does happen automatically, so you don't need to worry about that.
