### 2017-05-04 #242 @nicoguaro
@roguephysicist, you could check steps 5 - 7 from this [link](http://www.davidketcheson.info/2015/05/10/rock_solid_code.html). Basically, it is good to automate the testing and use something like ``nosetest``.

Regarding your tests this is the output I get

    Loading reference data.
    Calculating SHG yield for comparison.
    Traceback (most recent call last):

      File "<ipython-input-1-bc5c09d2d0d0>", line 1, in <module>
        runfile('D:/workspace/SHGYield/test/test.py', wdir='D:/workspace/SHGYield/test')

      File "C:\Anaconda2\lib\site-packages\spyder\utils\site\sitecustomize.py", line 866, in runfile
        execfile(filename, namespace)

      File "C:\Anaconda2\lib\site-packages\spyder\utils\site\sitecustomize.py", line 87, in execfile
        exec(compile(scripttext, filename, 'exec'), glob, loc)

      File "D:/workspace/SHGYield/test/test.py", line 13, in <module>
        test = np.loadtxt('test-output.dat', unpack=True)

I also think that you could add some details to the paper. Particularly, I would add the images.
      File "C:\Anaconda2\lib\site-packages\numpy\lib\npyio.py", line 858, in loadtxt
        fh = iter(open(fname, 'U'))

    IOError: [Errno 2] No such file or directory: 'test-output.dat'
