### 2017-05-04 #242 @roguephysicist
@nicoguaro, I will take a look and improve the tests. I will need to try the program on a windows machine to figure out the problem. So far, it has worked fine for me under Linux and macOS. I will hopefully be able to do this next week.

I will enhance the JOSS paper as well according to both your suggestions. Thanks!
