### 2017-05-05 9f7550c
removed FileNotFoundError, and changed math font in test
