### 2017-05-23 #242 @tacaswell
@tacaswell pinging my self so I get updates.  I will be a second reviewer on this.

I agree on automated tests.  It looks like you already have a good reference result (the hard part!), using `np.assert_almost_equal` instead of plotting should be easy.  I advise against `nose`, the authors have declared it effectively un-maintained, I suggest `pytest` instead.
