### 2017-05-25 #242 @roguephysicist
@nicoguaro @tacaswell, sorry for the absence -- work and long weekends have gotten in the way.

@tacaswell, Thanks for the suggestions -- I definitely agree that a numeric comparison is much simpler than plotting. I already implemented a small test using `pytest` and it was indeed very easy. I will produce a simple test suite encompassing the most important components of the program.

@nicoguaro, I will start working on a tutorial using a Jupyter notebook. It will walk the user through the different parts of the calculation and compare against the math featured in the references. It should be pretty informative for people looking to learn what this is all about.

I will hopefully be able to complete these tasks next week. Thanks again.
