### 2017-05-25 #242 @kyleniemeyer
@roguephysicist those plans sound great!
