### 2017-06-03 63f0953
Added self tests, housekeeping
