+++ shgyield.py
@@ -3,17 +3,14 @@
 shgyield.py is a python program designed to calculate the nonlinear reflection
 coefficient for semiconductor surfaces. It works in conjunction with the matrix
 elements calculated with ABINIT, an open source ab initio software, and TINIBA,
-our in-house optical calculation software.
+our in-house optical calculation software. For a complete overview of the
+theory, see PRB 94, 115314 (2016).
 
-For a complete overview of the theory, see PRB 94, 115314 (2016).
-
-Tested with Anaconda Python 4.0.0.
-
-requirements:
-sys, math, numpy, scipy
+required packages:
+`sys, yaml, numpy, scipy`
 
 usage:
-python shgyield.py <sample.in>
+`python shgyield.py input.yml`
 """
 
 import sys
@@ -62,7 +59,7 @@ def epsilon(in_file, norm):
     real = (data[1] + data[3] + data[5])/3      # real average
     imag = (data[2] + data[4] + data[6])/3      # imag average
     coma = real + 1j * imag                     # complex average
-    epsa = 1 + (4 * constants.pi * coma * norm) # epsilon with normalization
+    epsa = 1 + (4 * np.pi * coma * norm) # epsilon with normalization
     return epsa
 
 def shgload(infile):
@@ -76,24 +73,24 @@ def shgload(infile):
     chi2 = TINIBASCALE * PM2TOM2 * comp[:MAXE]
     return chi2
 
-def fresnel(kind, i, j, pol, freq):
+def freflc(pol, veci, vecj, epsi, epsj):
     '''
-    Generic fresnel factors, see Eq. (13) of PRB 94, 115314 (2016).
+    Generic reflection fresnel factors, see Eq. (13) of PRB 94, 115314 (2016).
     '''
-    wi = eval("w" + i + freq)
-    wj = eval("w" + j + freq)
-    epsi = eval("eps" + i + freq)
-    epsj = eval("eps" + j + freq)
-    if kind == "t":
-        if pol == "p":
-            factor = (2 * wi * np.sqrt(epsi * epsj))/(wi * epsj + wj * epsi)
-        elif pol == "s":
-            factor = (2 * wi)/(wi + wj)
-    elif kind == "r":
-        if pol == "p":
-            factor = ((wi * epsj) - (wj * epsi))/((wi * epsj) + (wj * epsi))
-        elif pol == "s":
-            factor = (wi - wj)/(wi + wj)
+    if pol == "p":
+        factor = ((veci * epsj) - (vecj * epsi))/((veci * epsj) + (vecj * epsi))
+    elif pol == "s":
+        factor = (veci - vecj)/(veci + vecj)
+    return factor
+
+def ftrans(pol, veci, vecj, epsi, epsj):
+    '''
+    Generic transmission fresnel factors, see Eq. (13) of PRB 94, 115314 (2016).
+    '''
+    if pol == "p":
+        factor = (2 * veci * np.sqrt(epsi * epsj))/(veci * epsj + vecj * epsi)
+    elif pol == "s":
+        factor = (2 * veci)/(veci + vecj)
     return factor
 
 def shgyield(gamma, riF):
@@ -101,8 +98,7 @@ def shgyield(gamma, riF):
[…]
