+++ tests/README.md
@@ -1,0 +1,8 @@
+Testing
+======================================
+
+Python requirements:
+`pytest`, `numpy`
+
+Usage:
+Simply run `pytest` in the root directory. It will automatically detect and run the appropriate tests contained in `self_test.py`. That script tests the different functions and calculations that are run in the main program. It will also compare the final result against a reference dataset in order to compare numerical accuracy.

[…]
