### 2017-06-05 767e4e5
second draft of paper
