### 2017-06-05 658cfb8
first draft of Jupyter tutorial, added readme for example
