+++ example/README.md
@@ -1,0 +1,6 @@
+Example and tutorial
+================================
+
+An example case using the Si(111)(1x1):H surface is included in this directory. Included is a very complete input file, `si111-input.yml`, all the components of the susceptibility tensors in `chi1-linear/` and `chi2-nonlinear/`, and some reference data with relevant plots.
+
+Also included is a Jupyter notebook, `sshg-tutorial.ipynb`, with a brief but thorough tutorial that explains some the of the theory and how the equations are coded. This tutorial walks the user through the battery of calculations required for the SSHG yield for this surface.
