+++ tests/README.md
@@ -5,4 +5,4 @@ Python requirements:
 `pytest`, `numpy`
 
 Usage:
-Simply run `pytest` in the root directory. It will automatically detect and run the appropriate tests contained in `self_test.py`. That script tests the different functions and calculations that are run in the main program. It will also compare the final result against a reference dataset in order to compare numerical accuracy.
+Simply run `pytest` in the root repo directory. It will automatically detect and run the appropriate tests contained in `self_test.py`. That script tests the different functions and calculations that are run in the main program. It will also compare the final result against a reference dataset in order to compare numerical accuracy.
