### 2017-06-05 80a2fac
updated figure paths in main readme
