### 2017-06-05 #242 @roguephysicist
Hi all,

I have finally finished the aforementioned additions to the program.

@nicoguaro, I added a Jupyter notebook tutorial in the `examples/` directory. I think it contains most of the relevant theory and code, but it was hard to keep it from getting too long with all the math.

@tacaswell, I added a test script in `tests/` that works with `pytest` under both MacOS and Linux (Windows pending). It runs 6 tests that try out different aspects of the program, ending with a numerical comparison against reference data. The run-time on my machine is around 0.50 seconds.

@kyleniemeyer, I added some content from the main README file to the JOSS paper. I looked at other examples, and kept it brief.

Anyway, looking forward to your comments and suggestions. Thanks again.
