### 2017-06-05 #242 @nicoguaro
@roguephysicist, I just ran your test under Windows, this is the output:

    ============================= test session starts =============================
    platform win32 -- Python 3.6.0, pytest-3.0.5, py-1.4.32, pluggy-0.4.0
    rootdir: C:\Users\nguarinz\workspace\SHGYield\tests, inifile:
    collected 0 items / 1 errors

    =================================== ERRORS ====================================
    ________________________ ERROR collecting self_test.py ________________________
    self_test.py:112: in <module>
        EPS = test_epsilon() # Creates the dictionary with the epsilons
    self_test.py:24: in test_epsilon
        epsl = read_eps('tests/data/SiH1x1-epsilon') # Epsilon from chi1, layered, normalized
    self_test.py:11: in read_eps
        data = np.loadtxt(in_file, unpack=True, skiprows=1)
    ..\..\..\Anaconda3\lib\site-packages\numpy\lib\npyio.py:805: in loadtxt
        fh = iter(open(fname))
    E   FileNotFoundError: [Errno 2] No such file or directory: 'tests/data/SiH1x1-epsilon'
    !!!!!!!!!!!!!!!!!!! Interrupted: 1 errors during collection !!!!!!!!!!!!!!!!!!!
    =========================== 1 error in 0.47 seconds ===========================

The rest seems to be fine!
