### 2017-06-05 #242 @roguephysicist
@nicoguaro, you have to run "pytest" in the `root/` directory, not in the `tests/` directory.
