### 2017-06-05 #242 @roguephysicist
Thanks @nicoguaro. I don't know why I said "`root/`," but I'm glad worked
anyway 😅.

On Jun 6, 2017 00:05, "Nicolás Guarín-Zapata" <notifications@github.com>
wrote:

@roguephysicist <https://github.com/roguephysicist>, you are right. It
works nicely now.
