### 2017-06-06 #242 @kyleniemeyer
@roguephysicist Your changes look great, between the paper, example Jupyter notebook, and addition of tests. Thanks! Could you now give the DOI for an archived version of the software, via (e.g.) Zenodo?

@nicoguaro for completeness, could you check off the review items at the top of the issue?

@arfon Do we need to update the version of the software in this issue? The software is now at v1.1, vs. v1.0.1 when submitted. (I can't remember if this is important or not...)
