### 2017-06-06 #242 @roguephysicist
Thanks @kyleniemeyer. I went ahead and added one more release with all these changes (v1.1.1). The Zenodo DOI for this release is `10.5281/zenodo.803483`. It does not resolve yet, but should be working in a few hours.

Let me know if anything else is needed on my part. Thanks again!
