### 2017-06-07 #242 @kyleniemeyer
@arfon ok, this one is good to go, other than the potential version change (which may not be necessary)!
