### 2017-06-07 #242 @roguephysicist
@arfon, I made an attempt at this, but I don't have any way to test it. Let me know if I need to do something different.
