### 2017-06-07 40f9457
Update paper.md
